Represent W as shift registers ISE recognises. BREAKS LOOP != 1. Worse on Altera.
For some reason, the Altera Quartus II synthesis tool doesn't
recognise this new code as shift registers, even though it kinda did
for the old code. This means it produces less efficient results with
this change than it did previously.

Also, this breaks values of LOOP != 1 (LOOP_LOG2 != 0). I need to
modify the code to handle this correctly.